Add files via upload
I uploaded the full thing where you can proper debug the code if something is wrong.
I rebuilt the code dividing everything into functions, where every scenario runs. Example: we have a Bedroom function, where everything concerning the bedroom happens. So it's easier for everyone to go on coding and for better readability.


What we need is a Player Class, a Zone Class and a Weapon Class, so, whenever the code runs, we fetch data from the database ( based on the name the user inserts, and upload them into the classes, so the player can proceed where he left his game